requirements.txt
albumentations==1.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cycler==0.11.0
docker-pycreds==0.4.0
easydict==1.13
efficientnet-pytorch==0.7.1
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.43
grad-cam==1.4.8
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
munch==4.0.0
networkx==2.6.3
numpy==1.21.6
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.0
Pillow==9.5.0
platformdirs==4.0.0
pretrainedmodels==0.7.4
protobuf==4.24.4
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.3.3
sentry-sdk==2.5.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
split-folders==0.5.1
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.4
ttach==0.0.3
typing_extensions==4.7.1
urllib3==1.26.11
wandb==0.17.1
zipp==3.15.0